test hashes

+++ lib/src/test/kotlin/sp/kx/storages/MockStreamerProvider.kt
@@ -3,7 +3,7 @@ package sp.kx.storages
 import java.util.UUID
 
 internal class MockStreamerProvider : SyncStreamsStorages.StreamerProvider {
-    override fun get(id: UUID, inputPointer: Long, outputPointer: Long): Streamer {
+    override fun get(id: UUID, inputPointer: Int, outputPointer: Int): Streamer {
         TODO("MockStreamerProvider:get($id, $inputPointer, $outputPointer)")
     }
 }
@@ -12,7 +12,7 @@ internal fun mockStreamerProvider(
     streamers: Map<UUID, Streamer>,
 ) : SyncStreamsStorages.StreamerProvider {
     return object : SyncStreamsStorages.StreamerProvider {
-        override fun get(id: UUID, inputPointer: Long, outputPointer: Long): Streamer {
+        override fun get(id: UUID, inputPointer: Int, outputPointer: Int): Streamer {
             return streamers[id] ?: TODO()
         }
     }
